+++ src/main/java/webserver/annotation/RequestMapping.java
@@ -1,4 +1,4 @@
-package annotation;
+package webserver.annotation;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
